requirements.txt
fastapi
uvicorn[standard]
kafka-python
clickhouse-connect
